Refactor glider email code, add deployment check after submission
Refactors the glider email code to move the main logic into the
glider_dac/glider_emails.py file.  Also attempts to handle differing
aggregate syntax in MongoDB which is used to find the list of
deployments to run the metadata compliance check against.
Adds a glider deployment compliance check after submission and sends an
email to the deployment provider when finished.