Fix compliance check scheduling
Fixes compliance check scheduling inside "on_complete" method of
deployments.  The on_complete method did not only run on completion, but
also ran whenever the data was "synced".  This update checks only if the
deployment is actually marked completed **and** there wasn't an existing
check run as determined by the presence of the "compliance_check_passed"
field in the deployments collection.  This fixes a problem where an
exorbitant number of email checks were attempted to run.